added multiple methodologies for 10mil gen + seed
