+++ database/models/room.js
@@ -27,4 +27,5 @@ const insertOne = (room, callback) => {
 module.exports = {
   findByID,
   insertOne,
+  RoomModel
 };
