--- database/seedFile/testSeed.csv
@@ -1,2 +1,0 @@
-id,sleepingArrangements
-1,[{"typeOfRoom":"Bedroom","furniture":{"typeOfFurniture":"single bed","qty":2}}]
